--- libs/ui/src/components/atoms/button/index.ts
@@ -1,1 +1,0 @@
-export * from './button';
