+++ libs/ui/src/components/atoms/dropdown/dropdown.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { BiError } from 'react-icons/bi';
-import { Button } from '../button/button';
+import { Button } from '../Button/Button';
 import { ParXs } from '../typography';
 import { DropdownItem, Dropdown } from './dropdown';
