+++ libs/ui/src/components/organisms/DaoHausConnect/DaoHausConnect.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { BiError } from 'react-icons/bi';
 import { RiUserAddLine } from 'react-icons/ri';
-import { ButtonProps, Button } from '../../atoms/button/button';
+import { ButtonProps, Button } from '../../atoms/Button/Button';
 import { Bold, ParMd, ParXs } from '../../atoms/typography';
 import styled from 'styled-components';
 import { violet } from '@radix-ui/colors';
